Changes in Server side client
I tried to make so that it tries to limit how many people are connected to the server. but I think it will be too complicated.

I made function were handle individuall cleints in sepearte threads so the application wont freeze,

I made a broadcast function

None of those function were tested.

I may discart the maximum number of players, but what I will do I will maintain dictionary so that colours are assigned to an address